Add files via upload
A Banking Management System (BMS) is a software application designed to manage and automate various banking activities, such as account management, transaction processing. The primary goal of a BMS is to provide a secure, efficient, and user-friendly platform for customers to interact with the bank, while also streamlining internal processes and reducing manual errors.